backend/requirements.txt
Django
djangorestframework
Pillow
tensorflow
opencv-python
easyocr
imutils
matplotlib